# Conflicts:
#	build.gradle
#	src/main/java/com/aor/hero/controller/game/HeroController.java
#	src/test/java/AttackDestructionTest.java
#	src/test/java/AttackPatternTest.java
#	src/test/java/GameTest.java
#	src/test/java/MenuTest.java
#	src/test/java/PositionTest.java
#	src/test/java/SoulTest.java

+++ src/main/java/com/aor/hero/model/game/elements/Hero.java
@@ -3,7 +3,7 @@ package com.aor.hero.model.game.elements;
 public class Hero extends Element {
     private int hp;
 
-    public Hero(double x, double y) {
+    public Hero(int x, int y) {
         super(x, y);
         this.hp = 30;
     }
